Added option to use hostname instead of GUID
Passing the option -n the nodes of the graph will contain the hostname
instead of the GUID.
Eg. from the line:
Ca	2 "H-123b456600dc1200"		# "host-x-y HCA-3"

The node will contain "host-x-y".
The HCA-3 part is also removed so that a host that is connected to
multiple switches will show this property in the resulting graph